Moving Test classes to the Testing namespace
