+++ BHoM/BHoM.csproj
@@ -195,9 +195,9 @@
     <Compile Include="Reflection\Debuging\EventType.cs" />
     <Compile Include="Reflection\Debuging\Log.cs" />
     <Compile Include="Reflection\Debuging\Event.cs" />
-    <Compile Include="Reflection\Testing\TestData.cs" />
-    <Compile Include="Reflection\Testing\TestResult.cs" />
-    <Compile Include="Reflection\Testing\UnitTest.cs" />
+    <Compile Include="Testing\TestData.cs" />
+    <Compile Include="Testing\TestResult.cs" />
+    <Compile Include="Testing\UnitTest.cs" />
     <Compile Include="Structural\Design\DesignGroup.cs" />
     <Compile Include="Structural\Design\StructuralLayout.cs" />
     <Compile Include="Structural\Design\IDesignable.cs" />
