+++ BHoM/Testing/UnitTest.cs
@@ -1,15 +1,14 @@
-﻿using System;
+﻿using BH.oM.Base;
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
-using System.Reflection;
-using BH.oM.Planning;
-using BH.oM.Base;
 
-namespace BH.oM.Reflection.Testing
+namespace BH.oM.Testing
 {
-    public class TestResult : BHoMObject
+    public class UnitTest : BHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
@@ -17,9 +16,8 @@ namespace BH.oM.Reflection.Testing
 
         public MethodBase Method { get; set; } = null;
 
-        public List<bool> Results { get; set; } = new List<bool>();
+        public List<TestData> Data { get; set; } = new List<TestData>();
 
-        public Issue Issue { get; set; } = null;
 
         /***************************************************/
     }
